backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
alembic==1.13.3
psycopg2
pydantic[email]
pillow==10.4.0
requests==2.31.0
httpx==0.27.0